requirements.txt
opencv-python~=4.11.0.86
numpy~=2.1.1
pillow~=11.1.0
torch~=2.6.0
torchvision~=0.21.0
matplotlib~=3.10.0
tqdm~=4.66.5
pandas~=2.2.3